# Experiment #143 in memory saving/performance in the Cell Collection