… tests/data/Cell/IndexesFromString.php