This is:

```
- [ ] a bugfix
- [ ] a new feature
- [X] refactoring
- [ ] additional unit tests
```

Checklist:

- [X] Changes are covered by unit tests
  - [X] Changes are covered by existing unit tests
  - [ ] New unit tests have been added
- [X] Code style is respected
- [X] Commit message explains **why** the change is made (see https://github.com/erlang/otp/wiki/Writing-good-commit-messages)
- [ ] CHANGELOG.md contains a short summary of the change
- [ ] Documentation is updated as necessary

### Why this change is needed?

Experiment [#143](https://github.com/PHPOffice/PhpSpreadsheet/pull/2791) in memory saving/performance in the Cell Collection.
Experiment [#132](https://github.com/PHPOffice/PhpSpreadsheet/pull/2786) reduced general memory usage of my big 500k cell test spreadsheet from 356,356KB down to 192,516KB, and Peak Memory Usage from 393,224K to 239624K; Experiment [#143](https://github.com/PHPOffice/PhpSpreadsheet/pull/2791) has reduced that Peak usage even further to just 208,904KB (although without any further improvement to general memory usage)

I'd already shaved an average of 2% off execution time for significant Cell Collection calls: now `getSortedCoordinates()` is down from 1.4s to 0.29s; `getHighestColumn()`/`getHighestRow()` without any argument down from around 0.7s to 0.6s, and with an argument from 0.55s to 0.1s. Cloning the cell collection reduced from 6.75s execution time to 6.3s, and using minimal additional memory to build the clone... it's harder to have any meaningful impact here because of the need to clone the cell object.
But I have also improved performance when deleting rows and columns: Initial timings for deleting a row with 256 columns: 0.72s => 0.125s; to delete a column with 2k rows: 0.58s => 0.115s

Two failures in the unit tests, related to the wrong Exception messages for invalid Named Range calls... both resolved by calling the validation check earlier in the `createNewCell()` method, failing faster.